Feature

+++ src/main/java/com/wip/bool/group/service/GroupMemberService.java
@@ -1,7 +1,7 @@
 package com.wip.bool.group.service;
 
-import com.wip.bool.cmmn.auth.AuthExecutor;
 import com.wip.bool.cmmn.type.OrderType;
+import com.wip.bool.exception.excp.AuthorizationException;
 import com.wip.bool.exception.excp.EntityNotFoundException;
 import com.wip.bool.exception.excp.ErrorCode;
 import com.wip.bool.group.domain.Group;
@@ -9,6 +9,7 @@ import com.wip.bool.group.domain.GroupMember;
 import com.wip.bool.group.domain.GroupMemberRepository;
 import com.wip.bool.group.domain.GroupRepository;
 import com.wip.bool.group.dto.GroupMemberDto;
+import com.wip.bool.user.domain.Role;
 import com.wip.bool.user.domain.User;
 import com.wip.bool.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -45,13 +46,14 @@ public class GroupMemberService {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException(userId, ErrorCode.NOT_FOUND_USER));
 
-        AuthExecutor<Long, GroupMember> authExecutor = new AuthExecutor<>();
+        Role role = user.getRole();
 
-        GroupMember groupMember = authExecutor.execute(user, groupMemberId,
-                gid -> groupMemberRepository.findById(gid)
-                        .orElseThrow(() -> new EntityNotFoundException(groupMemberId, ErrorCode.NOT_FOUND_GROUP)),
-                (uid, gid) -> groupMemberRepository.findById(uid, gid)
-                        .orElseThrow(() -> new EntityNotFoundException(groupMemberId, ErrorCode.NOT_FOUND_GROUP)));
+        if(role != Role.ROLE_ADMIN && role != Role.ROLE_NORMAL) {
+            throw new AuthorizationException();
+        }
+
+        GroupMember groupMember = groupMemberRepository.findById(userId, groupMemberId, role)
+                .orElseThrow(() -> new EntityNotFoundException(groupMemberId, ErrorCode.NOT_FOUND_GROUP));
 
         return groupMemberRepository.delete(groupMember);
     }
